+++ vite/vite.config.docs.mjs
@@ -54,5 +54,8 @@ export default defineConfig({
   },
   define: {
     'process.env.NODE_ENV': JSON.stringify('development')
+  },
+  optimizeDeps: {
+    include: ['y-websocket', 'y-indexeddb', 'y-protocols/awareness']
   }
 })